Self-escalation logic for linux/mac
